+++ src/main/java/net/silentchaos512/endertendril/data/DataGenerators.java
@@ -5,14 +5,17 @@ import net.minecraft.world.level.storage.loot.predicates.LootItemCondition;
 import net.neoforged.neoforge.common.data.GlobalLootModifierProvider;
 import net.neoforged.neoforge.data.event.GatherDataEvent;
 import net.silentchaos512.endertendril.EnderTendrilMod;
+import net.silentchaos512.endertendril.data.client.ModModelProvider;
+import net.silentchaos512.endertendril.data.tag.ModBlockTagsProvider;
+import net.silentchaos512.endertendril.data.tag.ModItemTagsProvider;
 import net.silentchaos512.endertendril.loot.ChestInjectorLootModifier;
+import net.silentchaos512.lib.data.recipe.LibRecipeProvider;
 
 public final class DataGenerators {
     private DataGenerators() {}
 
-    public static void gatherData(GatherDataEvent event) {
+    public static void gatherData(GatherDataEvent.Client event) {
         DataGenerator gen = event.getGenerator();
-        var existingFileHelper = event.getExistingFileHelper();
         var packOutput = gen.getPackOutput();
         var lookupProvider = event.getLookupProvider();
 
@@ -21,10 +24,9 @@ public final class DataGenerators {
         gen.addProvider(true, new ModItemTagsProvider(event, blocks));
 
         gen.addProvider(true, new ModLootTables(packOutput, lookupProvider));
-        gen.addProvider(true, new ModRecipesProvider(packOutput, lookupProvider));
+        gen.addProvider(true, LibRecipeProvider.createRunner(packOutput, lookupProvider, "Ender Tendril Recipes", ModRecipesProvider::new));
 
-        gen.addProvider(true, new ModBlockStateProvider(gen, existingFileHelper));
-        gen.addProvider(true, new ModItemModelProvider(gen, existingFileHelper));
+        gen.addProvider(true, new ModModelProvider(packOutput));
 
         gen.addProvider(true, new GlobalLootModifierProvider(packOutput, lookupProvider, EnderTendrilMod.MOD_ID) {
             @Override
@@ -32,5 +34,7 @@ public final class DataGenerators {
                 add("chest_loot_injector", new ChestInjectorLootModifier(new LootItemCondition[]{}));
             }
         });
+
+        gen.addProvider(true, new ModWorldGenGenerator(packOutput, lookupProvider).createRunner());
     }
 }