+++ src/main/java/net/silentchaos512/endertendril/data/tag/ModBlockTagsProvider.java
@@ -1,4 +1,4 @@
-package net.silentchaos512.endertendril.data;
+package net.silentchaos512.endertendril.data.tag;
 
 import net.minecraft.core.HolderLookup;
 import net.minecraft.tags.BlockTags;
@@ -10,7 +10,7 @@ import net.silentchaos512.endertendril.setup.ModTags;
 
 public class ModBlockTagsProvider extends BlockTagsProvider {
     public ModBlockTagsProvider(GatherDataEvent event) {
-        super(event.getGenerator().getPackOutput(), event.getLookupProvider(), EnderTendrilMod.MOD_ID, event.getExistingFileHelper());
+        super(event.getGenerator().getPackOutput(), event.getLookupProvider(), EnderTendrilMod.MOD_ID);
     }
 
     @Override