+++ src/main/java/net/silentchaos512/endertendril/item/EnderTendrilSeedItem.java
@@ -2,13 +2,14 @@ package net.silentchaos512.endertendril.item;
 
 import net.minecraft.core.BlockPos;
 import net.minecraft.core.Direction;
-import net.minecraft.world.item.ItemNameBlockItem;
+import net.minecraft.world.item.BlockItem;
+import net.minecraft.world.item.Item;
 import net.minecraft.world.item.context.BlockPlaceContext;
 import net.minecraft.world.level.block.Block;
 import net.minecraft.world.level.block.state.BlockState;
 
-public class EnderTendrilSeedItem extends ItemNameBlockItem {
-    public EnderTendrilSeedItem(Block blockIn, Properties properties) {
+public class EnderTendrilSeedItem extends BlockItem {
+    public EnderTendrilSeedItem(Block blockIn, Item.Properties properties) {
         super(blockIn, properties);
     }